+++ Software/Source/pijuice.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-__version__ = "1.9"
+__version__ = "1.10"
 
 import ctypes
 import functools